[Refactor] Use base fragment for fragments
