+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/CustomerFragment.kt
@@ -1,35 +1,21 @@
 package com.example.deliveryclonecoding.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.fragment.findNavController
 import com.example.deliveryclonecoding.R
 import com.example.deliveryclonecoding.databinding.FragmentCustomerBinding
+import com.example.deliveryclonecoding.ui.base.BaseFragment
 
-class CustomerFragment : Fragment() {
+class CustomerFragment : BaseFragment<FragmentCustomerBinding>(R.layout.fragment_customer) {
 
-    private var _binding: FragmentCustomerBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentCustomerBinding.inflate(inflater, container, false)
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
         setToolbarSettings()
         binding.layoutLogin.setOnClickListener {
             findNavController().navigate(R.id.action_customerFragment_to_loginFragment)
         }
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
     }
 
     private fun setToolbarSettings() {
