+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/FavoriteFragment.kt
@@ -1,27 +1,7 @@
 package com.example.deliveryclonecoding.ui
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import com.example.deliveryclonecoding.R
 import com.example.deliveryclonecoding.databinding.FragmentFavoriteBinding
+import com.example.deliveryclonecoding.ui.base.BaseFragment
 
-class FavoriteFragment : Fragment() {
-
-    private var _binding: FragmentFavoriteBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentFavoriteBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-}
+class FavoriteFragment : BaseFragment<FragmentFavoriteBinding>(R.layout.fragment_favorite)
