+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/LoginFragment.kt
@@ -1,9 +1,8 @@
 package com.example.deliveryclonecoding.ui
 
 import android.os.Bundle
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
+import com.example.deliveryclonecoding.R
 import com.example.deliveryclonecoding.databinding.FragmentLoginBinding
 import com.example.deliveryclonecoding.ui.base.BaseFragment
 import com.jakewharton.rxbinding3.view.clicks
@@ -12,23 +11,7 @@ import io.reactivex.Observable
 import io.reactivex.rxkotlin.addTo
 import java.util.concurrent.TimeUnit
 
-class LoginFragment : BaseFragment() {
-    private var _binding: FragmentLoginBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentLoginBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-
+class LoginFragment : BaseFragment<FragmentLoginBinding>(R.layout.fragment_login) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
