+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/OrderHistoryFragment.kt
@@ -1,27 +1,7 @@
 package com.example.deliveryclonecoding.ui
 
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
+import com.example.deliveryclonecoding.R
 import com.example.deliveryclonecoding.databinding.FragmentOrderHistoryBinding
+import com.example.deliveryclonecoding.ui.base.BaseFragment
 
-class OrderHistoryFragment : Fragment() {
-
-    private var _binding: FragmentOrderHistoryBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentOrderHistoryBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-}
+class OrderHistoryFragment : BaseFragment<FragmentOrderHistoryBinding>(R.layout.fragment_order_history)
