+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/SearchFragment.kt
@@ -1,27 +1,7 @@
 package com.example.deliveryclonecoding.ui
 
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
+import com.example.deliveryclonecoding.R
 import com.example.deliveryclonecoding.databinding.FragmentSearchBinding
+import com.example.deliveryclonecoding.ui.base.BaseFragment
 
-class SearchFragment : Fragment() {
-
-    private var _binding: FragmentSearchBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentSearchBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        _binding = null
-        super.onDestroyView()
-    }
-}
+class SearchFragment : BaseFragment<FragmentSearchBinding>(R.layout.fragment_search)
