+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/SplashFragment.kt
@@ -4,9 +4,7 @@ import android.os.Bundle
 import android.text.SpannableString
 import android.text.Spanned
 import android.text.style.AbsoluteSizeSpan
-import android.view.LayoutInflater
 import android.view.View
-import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
 import com.example.deliveryclonecoding.R
 import com.example.deliveryclonecoding.databinding.FragmentSplashBinding
@@ -16,24 +14,7 @@ import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.rxkotlin.addTo
 import java.util.concurrent.TimeUnit
 
-class SplashFragment : BaseFragment() {
-
-    private var _binding: FragmentSplashBinding? = null
-    private val binding by lazy { _binding!! }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        _binding = FragmentSplashBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
+class SplashFragment : BaseFragment<FragmentSplashBinding>(R.layout.fragment_splash) {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
