+++ Raemin/CloneCoding/app/src/main/java/com/example/deliveryclonecoding/ui/base/BaseFragment.kt
@@ -1,13 +1,34 @@
 package com.example.deliveryclonecoding.ui.base
 
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.annotation.LayoutRes
+import androidx.databinding.DataBindingUtil
+import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.Fragment
 import io.reactivex.disposables.CompositeDisposable
 
-abstract class BaseFragment : Fragment() {
+abstract class BaseFragment<VDB : ViewDataBinding>(
+    @LayoutRes private val layoutRes: Int
+) : Fragment() {
 
     protected val compositeDisposable = CompositeDisposable()
 
+    private var _binding: VDB? = null
+    protected val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
+        return binding.root
+    }
+
     override fun onDestroyView() {
+        _binding = null
         compositeDisposable.clear()
         super.onDestroyView()
     }
